requirements.txt
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
moviepy==1.0.3
numpy==1.20.1
Pillow==8.1.0
proglog==0.1.9
requests==2.25.1
tqdm==4.58.0
urllib3==1.26.3
youtube-dl==2021.2.22
